requirements.txt
secp256k1==0.13.2
flag==2.1.0
pyOpenSSL
pillow
pydub
gmpy2
sympy
galois


pwn
pycryptodome
tqdm
collections
timeout-decorator
pyzbar 
pillow